Split out ceph-graphite from ceph-grafana
