--- ansible/roles/ceph-grafana/files/cephmetrics-devel.repo
@@ -1,1 +1,0 @@
-../../../common/files/cephmetrics-devel.repo
