+++ ansible/roles/ceph-graphite/defaults/main.yml
@@ -1,0 +1,11 @@
+---
+defaults:
+  devel_packages:
+    yum:
+      - graphite-web
+      - python-carbon
+    apt:
+      - graphite-web
+      - apache2
+      - libapache2-mod-wsgi
+      - graphite-carbon
