+++ ansible/roles/ceph-graphite/meta/main.yml
@@ -1,0 +1,3 @@
+---
+dependencies:
+  - role: cephmetrics-common
