+++ ansible/roles/ceph-graphite/tasks/configure_carbon.yml
