+++ ansible/roles/ceph-graphite/tasks/configure_graphite.yml
