+++ ansible/roles/ceph-graphite/tasks/configure_graphite_web.yml
