+++ ansible/roles/ceph-graphite/tasks/install_packages.yml
@@ -1,0 +1,7 @@
+---
+- name: Install packages
+  package:
+    name: "{{ item }}"
+    state: latest
+  with_items: "{{ devel_packages[ansible_pkg_mgr] }}"
+  when: devel_mode
