+++ ansible/roles/ceph-graphite/tasks/merge_vars.yml
@@ -1,0 +1,1 @@
+../../cephmetrics-common/tasks/merge_vars.yml
