+++ ansible/roles/ceph-graphite/templates/graphite-api.socket
