+++ ansible/roles/ceph-graphite/templates/storage-schemas.conf
